Fix race condition in nrf_wifi_hal_dev_rem
Fix a race condition in nrf_wifi_hal_dev_rem where the PS deinit was
being called after doing a device removal from the bus. If the
PS timer expired after the device removal and before the PS deinit was
called it would try sending a WRSR2 command on the bus, resulting in a
hang. Move the PS deinit before bus device removal to fix this.

Signed-off-by: Sachin D Kulkarni <Sachin.Kulkarni@nordicsemi.no>