Fix Raw TX handling
Using a flag for per-packet handling is not ideal, esp. that flash is
cleared after TX, but in case of pending raw TX packets that will be
processed through TX done flag is unset and they won't be processed.

Move the check to per-packet and add APIs to set/get/check, this way
it's a clean implementation and handles sending from both TX and TX_DONE
paths.

This is a breaking change as the raw TX header is modified.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>